+++ EleFits/EleFits/impl/CompressionStrategy.hpp
@@ -11,8 +11,8 @@ namespace Euclid {
 namespace Fits {
 
 template <typename T>
-bool CompressAuto::apply(fitsfile* fptr, const ImageHdu::Initializer<T>& init) {
-
+bool CompressAuto::apply(fitsfile* fptr, const ImageHdu::Initializer<T>& init)
+{
   if (auto action = plio(init)) {
     return action->apply(fptr, init);
   }
@@ -29,21 +29,23 @@ bool CompressAuto::apply(fitsfile* fptr, const ImageHdu::Initializer<T>& init) {
 }
 
 template <typename T>
-std::unique_ptr<Compress<ShuffledGzip>> CompressAuto::gzip(const ImageHdu::Initializer<T>& init) {
-  return std::make_unique<Compress<ShuffledGzip>>(tiling(init), quantization<T>());
+std::unique_ptr<Compress<ShuffledGzip>> CompressAuto::gzip(const ImageHdu::Initializer<T>& init)
+{
+  return std::make_unique<Compress<ShuffledGzip>>(Tile::adaptive(), quantization<T>());
 }
 
 template <typename T>
-std::unique_ptr<Compress<Rice>> CompressAuto::rice(const ImageHdu::Initializer<T>& init) {
+std::unique_ptr<Compress<Rice>> CompressAuto::rice(const ImageHdu::Initializer<T>& init)
+{
   if (std::is_floating_point_v<T> && m_type == CompressionType::Lossless) {
     return nullptr;
   }
-  return std::make_unique<Compress<Rice>>(tiling(init), quantization<T>());
+  return std::make_unique<Compress<Rice>>(Tile::adaptive(), quantization<T>());
 }
 
 template <typename T>
-std::unique_ptr<Compress<HCompress>> CompressAuto::hcompress(const ImageHdu::Initializer<T>& init) {
-
+std::unique_ptr<Compress<HCompress>> CompressAuto::hcompress(const ImageHdu::Initializer<T>& init)
+{
   if (std::is_floating_point_v<T> && m_type == CompressionType::Lossless) {
     return nullptr;
   }
@@ -57,12 +59,12 @@ std::unique_ptr<Compress<HCompress>> CompressAuto::hcompress(const ImageHdu::Ini
   if (q.dithering() == Quantization::Dithering::NonZeroPixel) {
     q.dithering(Quantization::Dithering::EveryPixel);
   }
-  return std::make_unique<Compress<HCompress>>(hcompress_tiling(init), q, hcompress_scaling<T>());
+  return std::make_unique<Compress<HCompress>>(Tile::adaptive(), q, hcompress_scaling<T>());
 }
 
 template <typename T>
-std::unique_ptr<Compress<Plio>> CompressAuto::plio(const ImageHdu::Initializer<T>& init) {
-
+std::unique_ptr<Compress<Plio>> CompressAuto::plio(const ImageHdu::Initializer<T>& init)
+{
   constexpr auto bp = bitpix<T>();
 
   // Float or too large int
@@ -72,11 +74,12 @@ std::unique_ptr<Compress<Plio>> CompressAuto::plio(const ImageHdu::Initializer<T
   }
 
   // OK
-  return std::make_unique<Compress<Plio>>(tiling(init));
+  return std::make_unique<Compress<Plio>>();
 }
 
 template <typename T>
-Quantization CompressAuto::quantization() const {
+Quantization CompressAuto::quantization() const
+{
   if constexpr (std::is_integral_v<T>) {
     if (m_type != CompressionType::Lossy) {
       return Quantization(0);
@@ -93,47 +96,8 @@ Quantization CompressAuto::quantization() const {
 }
 
 template <typename T>
-Position<-1> CompressAuto::tiling(const ImageHdu::Initializer<T>& init) const {
-
-  static constexpr long min_size = 1024 * 1024;
-  if (shapeSize(init.shape) <= min_size) {
-    return Tile::whole();
-  }
-
-  const long row_width = init.shape[0];
-  const long row_size = row_width * sizeof(T);
-  const long row_count = min_size / row_size + 1;
-
-  // FIXME reach min_size using higher dimensions
-
-  return {row_width, row_count};
-}
-
-template <typename T>
-Position<-1> CompressAuto::hcompress_tiling(const ImageHdu::Initializer<T>& init) const {
-
-  const auto& shape = init.shape;
-
-  // Small image
-  if (shape[1] <= 30) {
-    // FIXME what about large rows?
-    return Tile::whole();
-  }
-
-  // Find acceptable row count
-  for (auto rows : {16, 24, 20, 30, 28, 26, 22, 18, 14}) {
-    auto modulo = shape[1] % rows;
-    if (modulo == 0 || modulo >= 4) {
-      return Tile::rowwise(rows);
-    }
-  }
-
-  // Fallback
-  return Tile::rowwise(17);
-}
-
-template <typename T>
-Scaling CompressAuto::hcompress_scaling() const {
+Scaling CompressAuto::hcompress_scaling() const
+{
   if constexpr (std::is_integral_v<T>) {
     if (m_type != CompressionType::Lossy) {
       return 0;