+++ EleFitsData/tests/src/Compression_test.cpp
@@ -14,14 +14,16 @@ BOOST_AUTO_TEST_SUITE(Compression_test)
 
 //-----------------------------------------------------------------------------
 
-BOOST_AUTO_TEST_CASE(disabled_scaling_test) {
+BOOST_AUTO_TEST_CASE(disabled_scaling_test)
+{
   Scaling scale = 0;
   BOOST_TEST(not scale);
   BOOST_TEST((scale.type() == Scaling::Type::Absolute));
   BOOST_TEST((scale.value() == 0));
 }
 
-BOOST_AUTO_TEST_CASE(absolute_scaling_test) {
+BOOST_AUTO_TEST_CASE(absolute_scaling_test)
+{
   Scaling scale = 8;
   BOOST_TEST(bool(scale));
   BOOST_TEST((scale.type() == Scaling::Type::Absolute));
@@ -29,7 +31,8 @@ BOOST_AUTO_TEST_CASE(absolute_scaling_test) {
   BOOST_CHECK_THROW((Scaling(-scale.value())), FitsError);
 }
 
-BOOST_AUTO_TEST_CASE(factor_scaling_test) {
+BOOST_AUTO_TEST_CASE(factor_scaling_test)
+{
   Scaling scale = Tile::rms * 2.5;
   BOOST_TEST(bool(scale));
   BOOST_TEST((scale.type() == Scaling::Type::Factor));
@@ -39,7 +42,8 @@ BOOST_AUTO_TEST_CASE(factor_scaling_test) {
   BOOST_TEST(identity.is_identity());
 }
 
-BOOST_AUTO_TEST_CASE(inverse_scaling_test) {
+BOOST_AUTO_TEST_CASE(inverse_scaling_test)
+{
   Scaling scale = Tile::rms / 4;
   BOOST_TEST(bool(scale));
   BOOST_TEST((scale.type() == Scaling::Type::Inverse));
@@ -49,7 +53,8 @@ BOOST_AUTO_TEST_CASE(inverse_scaling_test) {
   BOOST_TEST(identity.is_identity());
 }
 
-BOOST_AUTO_TEST_CASE(scaling_equality_test) {
+BOOST_AUTO_TEST_CASE(scaling_equality_test)
+{
   const auto a0 = Scaling(0);
   const auto a1 = Scaling(1);
   const auto f1 = Tile::rms * 1;
@@ -65,7 +70,8 @@ BOOST_AUTO_TEST_CASE(scaling_equality_test) {
   BOOST_TEST((f2 == i2));
 }
 
-BOOST_AUTO_TEST_CASE(default_quantization_test) {
+BOOST_AUTO_TEST_CASE(default_quantization_test)
+{
   Quantization q;
   BOOST_TEST(not q);
   BOOST_TEST(not q.level());
@@ -73,7 +79,8 @@ BOOST_AUTO_TEST_CASE(default_quantization_test) {
   BOOST_CHECK_THROW(q.dithering(Quantization::Dithering::EveryPixel), FitsError); // Cannot dither disabled q
 }
 
-BOOST_AUTO_TEST_CASE(default_dithering_test) {
+BOOST_AUTO_TEST_CASE(default_dithering_test)
+{
   Scaling level(4);
   Quantization q(level);
   BOOST_TEST(bool(q));
@@ -81,7 +88,8 @@ BOOST_AUTO_TEST_CASE(default_dithering_test) {
   BOOST_TEST((q.dithering() == Quantization::Dithering::EveryPixel));
 }
 
-BOOST_AUTO_TEST_CASE(quantization_dithering_test) {
+BOOST_AUTO_TEST_CASE(quantization_dithering_test)
+{
   const auto level = Tile::rms / 4; // CFITSIO default
   Quantization q(level);
   BOOST_TEST((q.level() == level));
@@ -91,7 +99,8 @@ BOOST_AUTO_TEST_CASE(quantization_dithering_test) {
   BOOST_TEST((q.dithering() == Quantization::Dithering::NonZeroPixel));
 }
 
-BOOST_AUTO_TEST_CASE(quantization_equality_test) {
+BOOST_AUTO_TEST_CASE(quantization_equality_test)
+{
   Quantization q0;
   Quantization q0n(0, Quantization::Dithering::None);
   Quantization q3(3);
@@ -108,8 +117,8 @@ BOOST_AUTO_TEST_CASE(quantization_equality_test) {
 }
 
 template <typename TAlgo>
-void test_algo_mixin_parameters(long dimension = 0) {
-
+void test_algo_mixin_parameters(long dimension = 0)
+{
   Position<-1> shape(dimension);
   std::fill(shape.begin(), shape.end(), 300);
   TAlgo algo(shape);
@@ -131,13 +140,14 @@ void test_algo_mixin_parameters(long dimension = 0) {
 
 // specific to the NoCompression algo
 template <>
-void test_algo_mixin_parameters<NoCompression>(long) {
+void test_algo_mixin_parameters<NoCompression>(long)
+{
   NoCompression algo;
-  BOOST_TEST((algo.tiling() == Position<-1>()));
+  BOOST_TEST((algo.tiling() == Position<-1> {0}));
 }
 
-BOOST_AUTO_TEST_CASE(algo_mixin_test) {
-
+BOOST_AUTO_TEST_CASE(algo_mixin_test)
+{
   test_algo_mixin_parameters<NoCompression>();
 
   for (long n = 0; n <= 6; ++n) {
@@ -159,8 +169,8 @@ BOOST_AUTO_TEST_CASE(algo_mixin_test) {
   }
 }
 
-BOOST_AUTO_TEST_CASE(hcompress_test) {
-
+BOOST_AUTO_TEST_CASE(hcompress_test)
+{
   const Position<-1>& shape {300, 200};
   HCompress algo(shape);